add seeders and factory

+++ database/migrations/2021_09_13_045519_create_checkout_items_table.php
@@ -18,7 +18,7 @@ class CreateCheckoutItemsTable extends Migration
             $table->foreignUuid('product_id')->constrained();
             $table->foreignUuid('checkout_id')->constrained();
             $table->decimal('item_price',13,7);
-            $table->decimal('item_quantity',13,7);
+            $table->integer('item_quantity');
             $table->timestamps();
         });
     }
